requirements.txt
django~=4.2.5
djangorestframework~=3.14.0
django-filter~=23.3
django-cors-headers~=4.2.0
djangorestframework-simplejwt~=5.3.0
drf-standardized-errors==0.12.5
psycopg2-binary==2.9.7
Pillow~=10.0.1
pyotp~=2.8.0
celery~=5.3.1
django-celery-beat~=2.5.0
django-celery-results~=2.5.1
django-templated-email~=3.0.1
django-templated-mail~=1.1.1
django-cities-light~=3.9.2
fcm-django~=2.0.0
django-environ~=0.11.2
django-ckeditor~=6.7.0
drf-spectacular~=0.26.5
drf-spectacular-sidecar~=2023.10.1
Jinja2~=3.1.2